requirements.txt
streamlit
requests
google-generativeai
langchain-google-genai
langchain-community
langchain